+++ packages/docz-core/src/commands/dev.ts
@@ -16,14 +16,20 @@ export const dev = async (args: Config) => {
   const env = envDotProp.get('node.env')
   const config = loadConfig(args)
   const port = await detectPort(config.port)
+  const hotPort = await detectPort(config.hotPort)
   const websocketPort = await detectPort(config.websocketPort)
   const entries = new Entries(config)
 
-  const bundler = webpack({ ...config, port }, env)
-  const server = await bundler.createServer(bundler.getConfig(env))
-  const { app } = await server.start()
+  envDotProp.set(
+    'webpack.server.overlay.ws.url',
+    `ws://${config.hotHost}:${hotPort}`
+  )
 
-  const newConfig = { ...config, websocketPort }
+  const newConfig = { ...config, websocketPort, hotPort, port }
+  const bundler = webpack(newConfig, env)
+  const bundlerConfig = await bundler.getConfig(env)
+  const server = await bundler.createServer(bundlerConfig)
+  const { app } = await server.start()
   const dataServer = new DataServer(
     app.server,
     config.websocketPort,