+++ packages/docz-theme-default/src/components/ui/Render/Handle.tsx
@@ -4,22 +4,25 @@ import styled, { css } from 'react-emotion'
 
 export const HANDLE_SIZE = '20px'
 
-const line = (color: string, position: string, horizontal: boolean) => css`
-  position: absolute;
-  display: block;
-  top: ${horizontal ? '50%' : position};
-  left: ${horizontal ? position : '50%'};
-  content: '';
-  width: ${horizontal ? '2px' : '25px'};
-  height: ${horizontal ? '25px' : '2px'};
-  background: ${color};
-  transform: translate(-50%, -50%);
-`
-
 interface HandleProps {
   full: boolean
   horizontal: boolean
+  theme?: any
 }
+
+const line = (position: string) => (p: HandleProps) => css`
+  content: '';
+  position: absolute;
+  display: block;
+  top: ${p.horizontal ? '50%' : position};
+  left: ${p.horizontal ? position : '50%'};
+  width: ${p.horizontal ? '2px' : '25px'};
+  height: ${p.horizontal ? '25px' : '2px'};
+  background: ${p.theme.mode === 'light'
+    ? darken(0.05, p.theme.colors.border)
+    : lighten(0.06, p.theme.colors.border)};
+  transform: translate(-50%, -50%);
+`
 
 const whenHorizontal = (on: any, off: any) => (p: HandleProps) =>
   p.horizontal ? on : off
@@ -33,12 +36,9 @@ export const Handle = styled('div')`
   display: block;
   width: ${whenHorizontal(HANDLE_SIZE, 'calc(100% + 5px)')};
   height: ${handleHeight};
-  border: 1px solid ${p => p.theme.colors.border};
+  border: 1px solid ${p => lighten(0.03, p.theme.colors.border)};
   border-radius: ${whenHorizontal('0 4px 4px 0', '0 0 4px 4px')};
-  background: ${p =>
-    p.theme.mode === 'light'
-      ? lighten(0.1, p.theme.colors.border)
-      : darken(0.1, p.theme.colors.border)};
+  background: ${p => darken(0.01, p.theme.colors.preBg)};
 
   ${whenHorizontal(
     `
@@ -52,10 +52,10 @@ export const Handle = styled('div')`
   )};
 
   &:after {
-    ${p => line(p.theme.colors.border, 'calc(50% + 3px)', p.horizontal)};
+    ${line('calc(50% + 3px)')};
   }
 
   &:before {
-    ${p => line(p.theme.colors.border, 'calc(50% - 3px)', p.horizontal)};
+    ${line('calc(50% - 3px)')};
   }
 `